% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy measures from arimab model}
\usage{
accuracy(object)
}
\arguments{
\item{object}{\code{arimab.predict} class object.}
}
\description{
The \code{accuracy()} returns the summary of accuracy measurements from a \code{arimab.predict} object.
}
\examples{
# data
data_i <- arimab::bacen
df_train <- data_i[1:100,]
df_test <- data_i[101:113,]

# model
fit_arima <- arimab::arimab(rate ~ arimab::lagb(inflation, -7),
 data = df_train)
df_test <- arimab::extractb(fit_arima, df_test)
pred_arima <- predict(fit_arima, df_test, h = 3)
arimab::accuracy(pred_arima)
#    n           me         mse      rmse       mae
# 1 93  0.009170212 0.007820629 0.1876783 0.1371477
# 2  3 -0.268580465 0.216406398 0.3284197 0.2685805
}
\references{
Hyndman, R. J., Athanasopoulos, G. (2016) \emph{Forecasting: principles and practice}. \url{https://otexts.com/fpp2}.
}
