% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{acfb}
\alias{acfb}
\title{Autocorrelation and Cross-Correlation Function Estimation}
\usage{
acfb(object, data = NULL, group = "acf",
     method = "prewhiten", ci = 0.95, ...)
acfb(object, data = NULL, group = "pacf")
acfb(object, data = NULL, group = "ccf")
}
\arguments{
\item{object}{A class object \code{formula} or a class object \code{arimab.model}. For \code{formula}, an object of class \code{formula} (or one that can be coerced to that class): a symbolic description of the model to be exploration. For \code{arimab.model}, a fit model with the function \code{arimab}.}

\item{data}{a class object of a \verb{data frame}.}

\item{group}{character class vector that indicates the estimates of the autocovariance or autocorrelation function.}

\item{method}{character class vector that indicates the method of transformation in the variables with option of \code{prewhiten} and \code{default} with ccf computes the cross-correlation or cross-covariance of two univariate series.}

\item{ci}{numeric class vector that indicates the confidence interval for \code{acf}.}

\item{...}{further arguments to be passed to \code{acf}, \code{pacf} or \code{ccf}.}
}
\value{
a class object \code{list.acfb} with time series exploration.
}
\description{
The \code{acfb()} computes estimates of the autocovariance or autocorrelation function with \code{acf}, \code{pacf} and \code{ccf}. Function \code{acfb} with group = 'pacf' is the function used for the partial autocorrelations. Function \code{acfb} with group = 'ccf' computes the cross-correlation or cross-covariance of two time series.
}
\details{
\code{acfb} is a wrapper of the \code{acf}, \code{pacf} and \code{ccf} function with the aim of simplifying an data explory for univariate and multiple time series. The focus is on fit per ⁠\verb{data frame}⁠. In addition to providing the \code{formula} object for data explory, it also makes the \code{list.acfb}, \code{list.pacfb} and \code{list.ccfb} object available for plotting.

An additional feature for the \code{acfb} function is the implementation of the pre-whitening method to analyze the cross-correlation for group = \code{ccf}. If the time series is stationary, the group = \code{default} method does not make any transformations to the time series.
}
\examples{
# exploration with data
plot(arimab::acfb(rate ~ ., arimab::bacen, group = "acf"))
plot(arimab::acfb(rate ~ ., arimab::bacen, group = "pacf"))
plot(arimab::acfb(rate ~ ., arimab::bacen, group = "ccf"))

# exploration with arimab
fit_arimab <- arimab::arimab(rate ~ inflation, arimab::bacen)
plot(arimab::acfb(fit_arimab))
plot(arimab::acfb(fit_arimab, group = "ccf"))
}
\references{
Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics with S}. Fourth Edition. Springer-Verlag.

Box, G., Jenkins, G. M., Reinsel, G. C. and Ljung, G. M (2015) \emph{Time series analysis: forecasting and control}. Fifth Edition. Wiley.
}
\seealso{
\code{\link{arimab}}
}
