% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{arimab}
\alias{arimab}
\title{Fitting for ARIMA time series modeling}
\usage{
arimab(formula, data, order = NULL, seasonal = NULL, 
       include_mean = FALSE, ...)
}
\arguments{
\item{formula}{character class vector that indicates the message of the title or subtitle of the comment.}

\item{data}{a class object of a \verb{data frame}.}

\item{order}{numeric class vector that indicates a specification of the non-seasonal part of the ARIMA model: the three integer components (p,d,q) are the AR order, the degree of differencing, and the MA order.}

\item{seasonal}{numeric class vector or list that indicates \code{order} and \code{period} to the a specification of the seasonal part of the ARIMA model, plus the period (which defaults to \code{frequency(x)}. This may be a list with components order and period, or just a numeric vector of length 3 which specifies the seasonal order. In the latter case the default period is used.}

\item{include_mean}{logical class vector that indicates the should the ARMA model include a mean/intercept term? The default is FALSE for time series.}

\item{...}{further arguments to be passed to \code{arima}.}
}
\value{
a class object \code{arimab.model}.
}
\description{
The \code{arimab()} is an adjustment with ARIMA and auto.ARIMA for univariate and multiple time series.
}
\details{
\code{arimab} is a wrapper of the \code{arima} function with the aim of simplifying an adjustment for univariate and multiple time series. The focus is on fit per \verb{data frame}. In addition to providing the \code{formula} object for model fitting, it also makes the \code{arimab.model} object available for plotting.

An additional feature for the \code{arimab} function is the availability of the \code{auto.arima} method if order = \code{NULL}. The implementation does not consider seasonality and other characteristics of the time series.
}
\examples{
# adjustment
fit_arima <- arimab::arimab(rate ~ ., arimab::bacen)
summary(fit_arima)
fit_arima <- update(fit_arima, rate ~ log(ibc_br))
summary(fit_arima)
arimab::diagnosticb(fit_arima)
plot(acfb(fit_arima))
plot(fit_arima)

# fitted 1
fit_arima_predict <- predict(fit_arima)
plot(fit_arima_predict)

# fitted 2
data_i <- arimab::bacen
df_train <- data_i[1:100,]
df_test <- data_i[101:113,]
fit_arima_uni  <- arimab::arimab(rate ~ 1, df_train)
fit_arima_mult <- arimab::arimab(rate ~ arimab::lagb(inflation, -7), df_train)
rbind(arimab::accuracy(predict(fit_arima_uni)),
     arimab::accuracy(predict(fit_arima_mult)))

# predict
select_vars <- arimab::extractb(fit_arima_mult, df_test)
fit_arima_predict <- predict(fit_arima_mult, 
     h = 4, newdata = select_vars)
plot(fit_arima_predict)
}
\references{
Brockwell, P. J. and Davis, R. A. (1996). \emph{Introduction to Time Series and Forecasting}. Springer, New York. Sections 3.3 and 8.3.

Durbin, J. and Koopman, S. J. (2001). \emph{Time Series Analysis by State Space Methods}. Oxford University Press.

Gardner, G, Harvey, A. C. and Phillips, G. D. A. (1980). Algorithm AS 154: An algorithm for exact maximum likelihood estimation of autoregressive-moving average models by means of Kalman filtering. \emph{Applied Statistics}, \bold{29}, 311--322. \ifelse{text}{doi:10.2307/2346910 <https://doi.org/10.2307/2346910>}{\ifelse{latex}{\href{https://doi.org/10.2307/2346910}{doi:10.2307\out{\slash{}}2346910}}{\href{https://doi.org/10.2307/2346910}{doi:10.2307/2346910}}}.

Harvey, A. C. (1993). \emph{Time Series Models}. 2nd Edition. Harvester Wheatsheaf. Sections 3.3 and 4.4.

Jones, R. H. (1980). Maximum likelihood fitting of ARMA models to time series with missing observations. \emph{Technometrics}, \bold{22}, 389--395. \ifelse{text}{doi:10.2307/1268324 <https://doi.org/10.2307/1268324>}{\ifelse{latex}{\href{https://doi.org/10.2307/1268324}{doi:10.2307\out{\slash{}}1268324}}{\href{https://doi.org/10.2307/1268324}{doi:10.2307/1268324}}}.

Ripley, B. D. (2002). \dQuote{Time series in \R 1.5.0}. \emph{R News}, \bold{2}(2), 2--7. \url{https://www.r-project.org/doc/Rnews/Rnews_2002-2.pdf}

Box, G., Jenkins, G. M., Reinsel, G. C. and Ljung, G. M (2015) \emph{Time series analysis: forecasting and control}. Fifth Edition. Wiley.

Hyndman, R. J., Athanasopoulos, G. (2016) \emph{Forecasting: principles and practice}. \url{https://otexts.com/fpp2}.

Cryer, J. Chan, K. (2008). \emph{Time Series Analysis: With Applications in R}. Springer New York.
}
\seealso{
\code{\link{lagb}}, \code{\link{acfb}}, \code{\link{diagnosticb}}
}
