% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacen}
\alias{bacen}
\title{Time Series dataset of the Central Bank of Brazil}
\format{
An object of class \code{data.frame} with 113 rows and 5 columns.
}
\source{
\url{https://www.bcb.gov.br/en}
}
\usage{
bacen
}
\description{
Set of Time Series data collected from the Central Bank of Brazil between 2014 and 2022 for the variables interest, inflation, exchange rate and ibc_br index.
}
\details{
\describe{
\item{date}{date with monthly periodicity}
\item{rate}{Average of the Copom's Selic target for interest rate}
\item{inflation}{Consumer price index in 12 months }
\item{currency}{Average nominal exchange rate in sales value}
\item{ibc_br}{percentage change for 12 months of the IBC-BR index of the Central Bank of Brazil}
}
}
\keyword{datasets}
