% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{diagnosticb}
\alias{diagnosticb}
\title{Diagnostics measures from a arimab model}
\usage{
diagnosticb(object, ci = 0.95)
}
\arguments{
\item{object}{\code{arimab.model} class object.}

\item{ci}{numeric class vector that indicates the confidence interval for \code{acf}.}
}
\description{
The \code{diagnosticb()} returns the summary of diagnostics measurements from a \code{arimab.model} object.
}
\details{
The method for arimab object plot residuals scaled by the estimate of their (individual) variance, and use the Ljung–Box version of the portmanteau test.
}
\examples{
data_i <- arimab::bacen
df_train <- data_i[1:100,]
fit_arima <- arimab::arimab(rate ~ arimab::lagb(inflation, -7),
 data = df_train)
arimab::diagnosticb(fit_arima)
plot(arimab::diagnosticb(fit_arima))
}
\references{
Box, G. E. P. and Pierce, D. A. (1970), Distribution of residual correlations in autoregressive-integrated moving average time series models. \emph{Journal of the American Statistical Association}, \bold{65}, 1509--1526.

Ljung, G. M. and Box, G. E. P. (1978), On a measure of lack of fit in time series models. \emph{Biometrika} \bold{65}, 297--303.

Harvey, A. C. (1993) \emph{Time Series Models}. 2nd Edition, Harvester Wheatsheaf, NY, pp.\sspace{}44, 45.
}
\seealso{
\code{\link{arimab}}
}
