% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extractb}
\alias{extractb}
\title{Extract Dataset from Fitted Model}
\usage{
extractb(object, newdata)
}
\arguments{
\item{object}{\code{arimab.model} class object.}

\item{newdata}{data frame for the test data.}
}
\description{
The \code{extractb()} is a tool for the \code{predict} function where \code{newdata != NULL} and serves to select variables from a test dataset from the model fitted in the training dataset.
}
\details{
The \code{extractb} function will be useful when there is transformation in the covariates for the fit with \code{arimab}. If the fit with \code{arimab} has past and future lags far above the number of observations in the test data frame, then it is possible that the output of the function is empty since it only generated missing values for the test data due to the transformations of the fit with \code{arimab}.
}
\examples{
# data
data_i <- arimab::bacen
df_train <- data_i[1:100,]
df_test <- data_i[101:113,]

# model
fit_arima <- arimab::arimab(rate ~ arimab::lagb(inflation, -7),
 data = df_train)
df_fit <- fit_arima[["data"]][["data"]]
df_newdata <- arimab::extractb(fit_arima, df_test)
colnames(df_fit) == colnames(df_newdata)
colnames(df_fit);colnames(df_newdata)

# predict
plot(predict(fit_arima, h = 3, newdata = df_newdata))
}
\seealso{
\code{\link{predict}}
}
