% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lagb}
\alias{lagb}
\title{Lag a Time Series}
\usage{
lagb(x, k)
}
\arguments{
\item{x}{numeric class vector.}

\item{k}{numeric class vector indicating the number of forward or backward lags.}
}
\description{
The \code{lagb()} moves a time series by shifting the vector in time backwards or forwards in a given number of observations.
}
\details{
\code{lagb} is a wrapper of the \code{lag} function with focus is on vector for \verb{data frame}. It is important to note that delays cause missing value in the vector, but in the function with \code{arimab} the missing values are removed for adjustment.
}
\examples{
library(arimab)

# explory data
rate_i <- head(arimab::bacen[["rate"]])
rate_i;lagb(rate_i, 0)
lagb(rate_i, -2)
lagb(rate_i, 2)

# fit model
data_i <- arimab::bacen
df_train <- data_i[1:100,]
fit_arima <- arimab::arimab(rate ~ lagb(inflation, -7), df_train)
summary(fit_arima)
plot(fit_arima)
}
\references{
Hyndman, R. J., Athanasopoulos, G. (2016) \emph{Forecasting: principles and practice}. \url{https://otexts.com/fpp2}.
}
\seealso{
\link{arimab}
}
