% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{univariateb}
\alias{univariateb}
\title{Autocorrelation and Partial Autocorrelation Function Estimation}
\usage{
univariateb(formula, data, d = NULL, ci = 0.95, ...)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced to that class): a symbolic description of the model to be exploration.s}

\item{data}{a class object of a \verb{data frame}.}

\item{d}{a numeric class vector indicating the amount of differentiation for the time series.}

\item{ci}{a numeric class vector that indicates the confidence interval for \code{acf}.}

\item{...}{further arguments to be passed to \code{acf}, \code{pacf} or \code{ccf}.}
}
\value{
a class object \code{arimab.univariateb} with time series exploration.
}
\description{
The \code{univariateb()} computes estimates of the autocorrelation and partial autocorrelation function with \code{acf} and \code{pacf}. Function \code{acfb} with group = 'pacf' is the function used for the partial autocorrelations to the univariate time series.
}
\details{
\code{univariateb} is a wrapper of the \code{acf} and \code{pacf} function with the aim of simplifying an data explory for univariate time series.
}
\examples{
# exploration with data
plot(arimab::univariateb(rate ~ 1, arimab::bacen))
plot(arimab::univariateb(inflation ~ 1, arimab::bacen))
plot(arimab::univariateb(inflation ~ 1, arimab::bacen, d = 2))
}
\references{
Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics with S}. Fourth Edition. Springer-Verlag.

Box, G., Jenkins, G. M., Reinsel, G. C. and Ljung, G. M (2015) \emph{Time series analysis: forecasting and control}. Fifth Edition. Wiley.
}
\seealso{
\code{\link{arimab}}
}
